Change RSS feed URL to point to Feedburner
Now that the blog is live, the RSS feed has a public URL. I like to use
Feedburner because it gives users a nicer page and gives me some
analytics. Also, the old blog was using Feedburner, so by changing its
origin feed to this new one, old subscribers will automatically be
subscribed to the new blog. Now that Feedburner is using the new feed
URL, I can replace it in the blog to show the Feedburner page instead of
the default raw XML one.