Fix reading time bug
The estimated reading time for a given post could return different
values depending if we were looking at a list of posts or the post page.
This was because in one situation Jekyll uses `post`, while in the other
it uses `page`. One of them doesn't convert the liquid tags in the
`category_and_date` include, which results in a different word count. So
passing the post's content as a parameter of that include fixes this.