Add portfolio to the landing page
Just like with the curriculum part, this also uses Jekyll's data helpers, which
let me express all my projects in a structured format and then just fetch them
with the appropriate styles in the landing page.

I didn't include the products of the companies I've worked on, because I think
a portfolio should only include projects where you feel you're the (or one of
the) main author/contributor. I might want to revisit this definition in the
future.