Add first post, which is a replacement for the old about page
Removes the markdown file for the old about page, and makes the first
blog post fill the same URL. If I want an about page in the future, I
can use the root path, instead of the blog's one.

This also adds a small style improvement to make images fit the expected
width (the cover image for the about post was larger).