Fix twitter card display
The image wasn't showing up on the twitter card because apparently it
needed to be bigger. This isn't documented anywhere.

Also removed duplicate metadata, as twitter can read from og metadata,
and changed the twitter card from summary_large_image, which doesn't
show a description and has a tiny title hidden in the image, to summary,
which has proper space for title and description.